requirements.txt
fastapi==0.79.0
uvicorn[standard]==0.18.2
python-socketio==4.6.1
loguru==0.6.0
pyyaml==6.0
invoke==1.7.1
requests==2.28.1
pyjson5==1.6.1
colorama==0.4.5
sqlalchemy==1.4.39
websocket-client==1.3.3
orjson==3.7.11